requirements.txt
assertpy==1.1
attrs==24.3.0
cachetools==5.5.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
distlib==0.3.9
exceptiongroup==1.2.2
execnet==2.1.1
filelock==3.16.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pyproject-api==1.8.0
PySocks==1.7.1
pytest==8.3.4
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-xdist==3.6.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.2.1
tox==4.23.2
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.28.1
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
